Add files via upload
PowerBI Dasboard Example.

The dataset contains the products listed on the website of online grocery store Big Basket.
Avenue Supermarts Limited, d/b/a DMart, is an Indian retail corporation that operates a chain of hypermarkets in India. 
It was founded by Radhakishan Damani in 2002, with its first branch in Powai's Hiranandani Gardens.